+++ src/app/app.module.ts
@@ -23,6 +23,7 @@ import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatButtonModule } from '@angular/material/button';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -30,6 +31,7 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { MatButtonToggleModule } from '@angular/material/button-toggle';
+import { MatRadioModule } from '@angular/material/radio';
 
 // import { MatAccordion } from '@angular/material/expansion';
 
@@ -40,6 +42,7 @@ import { CurrentUserService } from '@services/current-user.service';
 import { AuthenticationService } from '@services/authentication.service';
 import { EventsService } from '@services/events.service';
 import { UserService } from '@services/user.service';
+import { DisplayValuePipe } from '@pipes/display-value.pipe';
 import { SitesService } from '@services/sites.service';
 
 @NgModule({
@@ -51,6 +54,7 @@ import { SitesService } from '@services/sites.service';
         LoginComponent,
         RegistrationComponent,
         ConfirmComponent,
+        DisplayValuePipe,
     ],
     imports: [
         BrowserModule,
@@ -80,6 +84,7 @@ import { SitesService } from '@services/sites.service';
         MatInputModule,
         MatSelectModule,
         MatButtonToggleModule,
+        MatRadioModule,
         //MatAccordion
     ],
     providers: [
@@ -88,6 +93,8 @@ import { SitesService } from '@services/sites.service';
         EventsService,
         UserService,
         SitesService,
+        DisplayValuePipe,
+        MatSnackBar,
     ],
     bootstrap: [AppComponent],
     entryComponents: [