+++ src/app/app.settings.ts
@@ -10,66 +10,70 @@ export class APP_SETTINGS {
         return environment.api_root;
     }
 
+    /* istanbul ignore next */
     public static set environment(env: string) {
         this._environment = env;
     }
     public static get IS_LOGGEDIN(): boolean {
         return (
-            !!sessionStorage.getItem('username') &&
-            !!sessionStorage.getItem('password')
+            !!localStorage.getItem('username') &&
+            !!localStorage.getItem('password')
         );
-    }
-    public static get API_USERNAME(): string {
-        return 'admin';
-    }
-    public static get API_PASSWORD(): string {
-        return 'icejamsadmin';
-    }
-
-    public static get DEFAULT_COUNTRY(): string {
-        return 'USA';
     }
 
     public static get AUTH_URL(): string {
         return this.API_ROOT + 'login';
     }
 
+    /* istanbul ignore next */
     public static get AGENCIES_URL(): string {
         return this.API_ROOT + 'agencies/';
     }
+    /* istanbul ignore next */
     public static get DAMAGE_TYPES_URL(): string {
         return this.API_ROOT + 'damagetypes/';
     }
+    /* istanbul ignore next */
     public static get EVENTS_URL(): string {
         return this.API_ROOT + 'sitevisits/';
-    } // i.e ice-jams
+    }
+    /* istanbul ignore next */
     public static get FILE_TYPES_URL(): string {
         return this.API_ROOT + 'filetypes/';
     }
+    /* istanbul ignore next */
     public static get ICE_CONDITION_TYPES_URL(): string {
         return this.API_ROOT + 'iceconditiontypes/';
     }
+    /* istanbul ignore next */
     public static get JAM_TYPES_URL(): string {
         return this.API_ROOT + 'jamtypes/';
     }
+    /* istanbul ignore next */
     public static get OBSERVERS_URL(): string {
         return this.API_ROOT + 'observers/';
     }
+    /* istanbul ignore next */
     public static get STAGES_URL(): string {
         return this.API_ROOT + 'stagetypes/';
     }
+    /* istanbul ignore next */
     public static get ROLES_URL(): string {
         return this.API_ROOT + 'roles/';
     }
+    /* istanbul ignore next */
     public static get ROUGHNESS_TYPES_URL(): string {
         return this.API_ROOT + 'roughnesstypes/';
     }
+    /* istanbul ignore next */
     public static get RIVER_CONDITION_TYPES(): string {
         return this.API_ROOT + 'riverconditiontypes/';
     }
+    /* istanbul ignore next */
     public static get SITES_URL(): string {
         return this.API_ROOT + 'sites';
     }
+    /* istanbul ignore next */
     public static get WEATHER_CONDITION_TYPES_URL(): string {
         return this.API_ROOT + 'weatherconditiontypes/';
     }
@@ -92,7 +96,7 @@ export class APP_SETTINGS {
         return new HttpHeaders({
             Authorization:
                 'Basic ' +
-                btoa(sessionStorage.username + ':' + sessionStorage.password),
+                btoa(localStorage.username + ':' + localStorage.password),
             Accept: 'application/json',
         });
     }