+++ src/app/home/home.component.spec.ts
@@ -2,13 +2,20 @@ import { HttpClient, HttpHandler } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CurrentUserService } from '@services/current-user.service';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatSnackBar } from '@angular/material/snack-bar';
 // import { by } from '@angular/platform-browser';
-import * as L from 'leaflet';
+
+declare let L: any;
+import 'leaflet';
+import 'leaflet-draw';
 
 import { Event } from '@interfaces/event';
 
 import { HomeComponent } from './home.component';
 import { APP_UTILITIES } from '@app/app.utilities';
+import { APP_SETTINGS } from '@app/app.settings';
+import { MAP_CONSTANTS } from './map-constants';
+import { DisplayValuePipe } from '@app/pipes/display-value.pipe';
 
 describe('HomeComponent', () => {
     let component: HomeComponent;
@@ -23,6 +30,8 @@ describe('HomeComponent', () => {
                 CurrentUserService,
                 HttpClient,
                 HttpHandler,
+                DisplayValuePipe,
+                MatSnackBar,
             ],
         }).compileComponents();
         component = TestBed.inject(HomeComponent);
@@ -90,4 +99,62 @@ describe('HomeComponent', () => {
         ];
         expect(response).toEqual(expectedResponse);
     });
+
+    it('currentFilter is set to user-stored filter if it exists', () => {
+        localStorage.setItem('currentFilter', '[1, 7, 8, 15]');
+        component.setCurrentFilter();
+        expect(component.currentFilter).toEqual(JSON.parse('[1, 7, 8, 15]'));
+    });
+
+    it('currentFilter is set to default if no storage filter set', () => {
+        localStorage.clear();
+        component.setCurrentFilter();
+        expect(component.currentFilter).toEqual(
+            APP_SETTINGS.DEFAULT_FILTER_QUERY
+        );
+    });
+
+    it('#getDrawnItemPopupContent returns the appropriate content response', () => {
+        // component.create
+        let latlngs = [
+            [37, -109.05],
+            [41, -109.03],
+            [41, -102.05],
+            [37, -102.04],
+        ];
+        var polygon = L.polygon(latlngs, { color: 'red' }).addTo(component.map);
+        let polygonResponse = component.getDrawnItemPopupContent(polygon);
+        expect(polygonResponse).toEqual(jasmine.any(String));
+        expect(polygonResponse).toContain('Area: ');
+
+        var line = L.polyline(latlngs, { color: 'blue' }).addTo(component.map);
+        let lineResponse = component.getDrawnItemPopupContent(line);
+        expect(lineResponse).toEqual(jasmine.any(String));
+        expect(lineResponse).toContain('Distance: ');
+
+        var notLine = L.polyline([latlngs[0]], { color: 'blue' }).addTo(
+            component.map
+        );
+        let notLineResponse = component.getDrawnItemPopupContent(notLine);
+        expect(notLineResponse).toEqual(jasmine.any(String));
+        expect(notLineResponse).toContain('Distance: N/A');
+
+        let notLayer = L.popup();
+        let notLayerResponse = component.getDrawnItemPopupContent(notLayer);
+        expect(notLayerResponse).toBeNull();
+    });
+
+    it('#eventFocus sets map to event focused view', () => {
+        // temporarily sets map to U.S, extent instead of event's extent
+        // first set the view to somehting not default to test that the update works
+        let notDefaultCenter = new L.LatLng(55.8283, -125.5795);
+        component.map.setView(notDefaultCenter, 9);
+        component.eventFocus();
+        let mapCenter = component.map.getCenter();
+        let mapZoom = component.map.getZoom();
+        expect(mapCenter).toEqual(MAP_CONSTANTS.defaultCenter);
+        expect(mapZoom).toEqual(MAP_CONSTANTS.defaultZoom);
+    });
+
+    xit('#displayState', () => {});
 });