+++ src/karma.conf.js
@@ -17,8 +17,14 @@ module.exports = function (config) {
         },
         coverageIstanbulReporter: {
             dir: require('path').join(__dirname, '../coverage'),
-            reports: ['html', 'lcovonly'],
+            reports: ['html', 'lcovonly', 'json-summary'],
             fixWebpackSourcePaths: true,
+            thresholds: {
+                statements: 80,
+                lines: 80,
+                branches: 50,
+                functions: 80,
+            },
         },
         reporters: ['progress', 'kjhtml'],
         port: 9876,